chore: remove recorded files and recording ui
