+++ meteor/client/ui/TestTools/index.tsx
@@ -4,9 +4,7 @@ import { withTranslation } from 'react-i18next'
 import { Translated } from '../../lib/ReactMeteorData/react-meteor-data'
 import { Route, Switch, Redirect, NavLink } from 'react-router-dom'
 
-import { RecordingsList, RecordingsStudioSelect } from './RecordingsList'
 import { TimelineView, TimelineStudioSelect } from './Timeline'
-import { RecordingView } from './RecordingView'
 import { UserLogPlayerPage, UserLogRundownSelect } from './UserLogPlayer'
 import { MeteorReactComponent } from '../../lib/MeteorReactComponent'
 import { PubSub } from '../../../lib/api/pubsub'
@@ -22,12 +20,6 @@ const StatusMenu = withTranslation()(
 
 			return (
 				<div className="tight-xs htight-xs text-s">
-					<NavLink
-						activeClassName="selectable-selected"
-						className="testTools-menu__testTools-menu-item selectable clickable"
-						to={'/testTools/recordings'}>
-						<h3>{t('Recordings')}</h3>
-					</NavLink>
 					<NavLink
 						activeClassName="selectable-selected"
 						className="testTools-menu__testTools-menu-item selectable clickable"
@@ -77,12 +69,8 @@ class Status extends MeteorReactComponent<Translated<IStatusProps>> {
 							<Switch>
 								<Route path="/testTools/timeline/:studioId" component={TimelineView} />
 								<Route path="/testTools/timeline" component={TimelineStudioSelect} />
-								<Route path="/testTools/recordings/:studioId/:recordingId" component={RecordingView} />
-								<Route path="/testTools/recordings/:studioId" component={RecordingsList} />
-								<Route path="/testTools/recordings" component={RecordingsStudioSelect} />
 								<Route path="/testTools/userlogplayer/:rundownPlaylistId" component={UserLogPlayerPage} />
 								<Route path="/testTools/userlogplayer" component={UserLogRundownSelect} />
-								<Redirect to="/testTools/recordings" />
 							</Switch>
 						</div>
 					</div>
