+++ meteor/lib/collections/Timeline.ts
@@ -42,16 +42,11 @@ export interface TimelineObjGeneric extends TimelineObjectCoreExt {
 export enum TimelineObjType {
 	/** Objects played in a rundown */
 	RUNDOWN = 'rundown',
-	/** Objects controlling recording */
-	RECORDING = 'record',
 	/** Objects controlling manual playback */
 	MANUAL = 'manual',
 }
 export interface TimelineObjRundown extends TimelineObjGeneric {
 	objectType: TimelineObjType.RUNDOWN
-}
-export interface TimelineObjRecording extends TimelineObjGeneric {
-	objectType: TimelineObjType.RECORDING
 }
 export interface TimelineObjManual extends TimelineObjGeneric {
 	objectType: TimelineObjType.MANUAL
