+++ meteor/lib/main.ts
@@ -19,7 +19,6 @@ import './collections/PeripheralDeviceCommands'
 import './collections/PeripheralDevices'
 import './collections/PieceInstances'
 import './collections/Pieces'
-import './collections/RecordedFiles'
 import './collections/RundownBaselineAdLibPieces'
 import './collections/RundownBaselineObjs'
 import './collections/RundownLayouts'
