+++ meteor/server/DatabaseCaches.ts
@@ -21,7 +21,6 @@ import { PieceInstance, PieceInstances } from '../lib/collections/PieceInstances
 import { Studio, Studios, StudioId } from '../lib/collections/Studios'
 import { Timeline, TimelineObjGeneric, TimelineComplete } from '../lib/collections/Timeline'
 import { RundownBaselineObj, RundownBaselineObjs } from '../lib/collections/RundownBaselineObjs'
-import { RecordedFile, RecordedFiles } from '../lib/collections/RecordedFiles'
 import { PeripheralDevice, PeripheralDevices } from '../lib/collections/PeripheralDevices'
 import {
 	protectString,
@@ -140,7 +139,6 @@ export class CacheForStudioBase extends Cache {
 	RundownPlaylists: DbCacheWriteCollection<RundownPlaylist, DBRundownPlaylist>
 	// Studios: DbCacheWriteCollection<Studio, Studio>
 	Timeline: DbCacheWriteCollection<TimelineComplete, TimelineComplete>
-	RecordedFiles: DbCacheWriteCollection<RecordedFile, RecordedFile>
 
 	constructor(studioId: StudioId) {
 		super()
@@ -149,7 +147,6 @@ export class CacheForStudioBase extends Cache {
 		this.RundownPlaylists = new DbCacheWriteCollection<RundownPlaylist, DBRundownPlaylist>(RundownPlaylists)
 		// this.Studios = new DbCacheWriteCollection<Studio, Studio>(Studios)
 		this.Timeline = new DbCacheWriteCollection<TimelineComplete, TimelineComplete>(Timeline)
-		this.RecordedFiles = new DbCacheWriteCollection<RecordedFile, RecordedFile>(RecordedFiles)
 	}
 	defer(fcn: DeferredFunction<CacheForStudioBase>) {
 		return super.defer(fcn)
@@ -187,7 +184,6 @@ async function fillCacheForStudioBaseWithData(
 	await Promise.all([
 		makePromise(() => cache.RundownPlaylists.prepareInit({ studioId: studioId }, initializeImmediately)),
 		makePromise(() => cache.Timeline.prepareInit({ _id: studioId }, initializeImmediately)),
-		makePromise(() => cache.RecordedFiles.prepareInit({ studioId: studioId }, initializeImmediately)),
 	])
 
 	return cache
