+++ meteor/server/api/system.ts
@@ -31,7 +31,6 @@ import { Parts } from '../../lib/collections/Parts'
 import { PeripheralDeviceCommands } from '../../lib/collections/PeripheralDeviceCommands'
 import { PeripheralDevices, PeripheralDeviceId } from '../../lib/collections/PeripheralDevices'
 import { Pieces } from '../../lib/collections/Pieces'
-import { RecordedFiles } from '../../lib/collections/RecordedFiles'
 import { RundownBaselineAdLibActions } from '../../lib/collections/RundownBaselineAdLibActions'
 import { RundownBaselineAdLibPieces } from '../../lib/collections/RundownBaselineAdLibPieces'
 import { RundownBaselineObjs } from '../../lib/collections/RundownBaselineObjs'
@@ -359,10 +358,6 @@ function cleanupOldDataInner(actuallyCleanup: boolean = false): CollectionCleanu
 		removeByQuery('PieceInstances', PieceInstances, {
 			rundownId: { $nin: rundownIds },
 		})
-	}
-	// RecordedFiles
-	{
-		results.push(ownedByStudioId('RecordedFiles', RecordedFiles))
 	}
 	// RundownBaselineAdLibActions
 	{
