+++ meteor/server/api/userActions.ts
@@ -15,8 +15,6 @@ import { SourceLayerType, IngestPart, IngestAdlib, ActionUserData } from 'tv-aut
 import { storeRundownPlaylistSnapshot } from './snapshot'
 import { registerClassToMeteorMethods } from '../methods'
 import { ServerRundownAPI } from './rundown'
-import { ServerTestToolsAPI, getStudioConfig } from './testTools'
-import { RecordedFiles, RecordedFileId } from '../../lib/collections/RecordedFiles'
 import { saveEvaluation } from './evaluations'
 import { MediaManagerAPI } from './mediaManager'
 import { IngestDataCache, IngestCacheType } from '../../lib/collections/IngestDataCache'
@@ -595,47 +593,6 @@ export function resyncSegment(context: MethodContext, rundownId: RundownId, segm
 
 	return ClientAPI.responseSuccess(ServerRundownAPI.resyncSegment(context, segment.rundownId, segmentId))
 }
-export function recordStop(context: MethodContext, studioId: StudioId) {
-	check(studioId, String)
-	StudioContentWriteAccess.recordedFiles(context, studioId)
-
-	const record = RecordedFiles.findOne({
-		studioId: studioId,
-		stoppedAt: { $exists: false },
-	})
-	if (!record) return ClientAPI.responseError(`No active recording for "${studioId}" was found!`)
-	return ClientAPI.responseSuccess(ServerTestToolsAPI.recordStop(context, studioId))
-}
-
-export function recordStart(context: MethodContext, studioId: StudioId, fileName: string) {
-	check(studioId, String)
-	check(fileName, String)
-
-	const access = StudioContentWriteAccess.recordedFiles(context, studioId)
-	const studio = access.studio
-	if (!studio) return ClientAPI.responseError(`Studio "${studioId}" not found!`)
-
-	const active = RecordedFiles.findOne({
-		studioId: studioId,
-		stoppedAt: { $exists: false },
-	})
-	if (active) return ClientAPI.responseError(`There is already an active recording in studio "${studioId}"!`)
-
-	const config = getStudioConfig(studio)
-	if (!config.recordings.channelIndex)
-		return ClientAPI.responseError(`Cannot start recording due to a missing setting: "channel".`)
-	if (!config.recordings.deviceId)
-		return ClientAPI.responseError(`Cannot start recording due to a missing setting: "device".`)
-	if (!config.recordings.decklinkDevice)
-		return ClientAPI.responseError(`Cannot start recording due to a missing setting: "decklink".`)
-	if (!config.recordings.channelIndex)
-		return ClientAPI.responseError(`Cannot start recording due to a missing setting: "channel".`)
-
-	return ClientAPI.responseSuccess(ServerTestToolsAPI.recordStart(context, studioId, fileName))
-}
-export function recordDelete(context: MethodContext, fileId: RecordedFileId) {
-	return ClientAPI.responseSuccess(ServerTestToolsAPI.recordDelete(context, fileId))
-}
 export function mediaRestartWorkflow(context: MethodContext, workflowId: MediaWorkFlowId) {
 	return ClientAPI.responseSuccess(MediaManagerAPI.restartWorkflow(context, workflowId))
 }
@@ -1017,15 +974,6 @@ class ServerUserActionAPI extends MethodContextAPI implements NewUserActionAPI {
 	}
 	resyncSegment(_userEvent: string, rundownId: RundownId, segmentId: SegmentId) {
 		return traceAction('userAction.resyncSegment', resyncSegment, this, rundownId, segmentId)
-	}
-	recordStop(_userEvent: string, studioId: StudioId) {
-		return makePromise(() => recordStop(this, studioId))
-	}
-	recordStart(_userEvent: string, studioId: StudioId, name: string) {
-		return makePromise(() => recordStart(this, studioId, name))
-	}
-	recordDelete(_userEvent: string, id: RecordedFileId) {
-		return makePromise(() => recordDelete(this, id))
 	}
 	mediaRestartWorkflow(_userEvent: string, workflowId: MediaWorkFlowId) {
 		return makePromise(() => mediaRestartWorkflow(this, workflowId))
