+++ meteor/server/main.ts
@@ -24,7 +24,6 @@ import './api/showStyles'
 import './api/snapshot'
 import './api/studios'
 import './api/system'
-import './api/testTools'
 import './api/userActions'
 import './methods'
 import './migration/api'
